barcode_scanner: add video buffer check
An issue was found that video buffer dequeued may not be written by
ISI. This commit checks the error by writing a magic number to first
pixel before enqueue and check if it has changed after dequeue. Video
buffers that has the magic number unchanged are treated invalid. Error
log is be printed and the buffer is enqueued immediately.

Signed-off-by: Chekhov Ma <chekhov.ma@nxp.com>